requirements.txt
beautifulsoup4~=4.13.0
requests~=2.32.3
tabulate~=0.9.0
click~=8.1.8
python-dotenv~=1.0.1